+++ Swipbox_VIPER/Swipbox_VIPER/Core/Extensions/UIKit/UIView.swift
@@ -1,5 +1,5 @@
 //
-//  UIKit Extention.swift
+//  UIView.swift
 //  Swipbox_VIPER
 //
 //  Created by Muneeb Ali on 05/10/2024.
@@ -7,22 +7,6 @@
 
 import UIKit
 
-extension UIScreen {
-    static var screenWidth: CGFloat {
-        return UIScreen.main.bounds.size.width
-    }
-    
-    static var screenHeight: CGFloat {
-        return UIScreen.main.bounds.size.height
-    }
-    
-    static var screenDiagonal: CGFloat {
-        return sqrt(screenWidth * screenWidth + screenHeight * screenHeight)
-    }
-}
-
-
-//MARK: - UIView
 extension UIView {
     func applyShadow( color: UIColor = .black, offset: CGSize = CGSize(width: 0, height: 3), opacity: Float = 0.4, radius: CGFloat = 5.0) {
         layer.shadowColor = color.cgColor
