+++ Swipbox_VIPER/Swipbox_VIPER/Presentation/Base/BaseStructures.swift
@@ -1,0 +1,15 @@
+//
+//  BaseStructures.swift
+//  Swipbox_VIPER
+//
+//  Created by Muneeb Ali on 10/10/2024.
+//
+
+import Foundation
+
+// Constants for device dimensions (example: iPhone 12)
+struct ScreenSize {
+    static let baseWidth: CGFloat = 390
+    static let baseHeight: CGFloat = 844
+    static let baseDiagonal: CGFloat = 877
+}
